invalid session and unstable connections
In particular, this solves an issue where the underlying client
connection would simply rotate over the list of server forever
if the client has an invalid session. This causes issue
when running, for instance, a single ZK node, and restarting
it while a MAGE cluster is connected to it (all nodes end up
spamming the ZooKeeper instance once it comes back).

Fixes #108