processManager: mage worker id and || fix
Parallelism for worker restart was set to 4, and never changed;
instead, we now tell to start all workers at once on start, and
to reload them one by one on reload.

Unique ID per worker; if a worker dies and is replaced,
the replaced worker will have the same `mage.workerId`,
thus allowing for more consistent logging and metrics.