mmrp: relay map corruption
MMRP relies on service discovery to keep its list of known relays
up to date. However, under certain conditions, and depending on the
behavior of the service discovery backend in use, you may end up
with two relays sharing the same URI. This, in and of itself, appeared
to be the cause of multiple issues.

You may also end up with case where lingering nodes are cleaned up
after the replacing node has been added; in such cases, a disconnect was
still triggered, thus leaving the MMRP cluster in an inconsistent state
where messages where silently black-holed.

To solve this problem, this commit:

  1. On relay down, disconnect based on the cluster ID instead
     of disconnecting based on the uri (we still verify that the URI
     is consistent)
  2. On relay up, disconnect any relay with the same URI but a different
     clusterId