serviceDiscovery: do not mess with the IP order
Changing the IP order breaks how clustering works on Docker Swarm.

This is due to the fact that interfaces have *two* IP addresses:

  - One being the IP being used for communication on the overlay network
  - One being used locally for local routing on the local swarm node

Connecting to that second interface results in MMRP and
other services relying on service discovery to essentially connect
to themselves.

I also took the opportunity to uniformize how Zookeeper and Consul
work (or at least how they select the IP to use, and allowing
similar configuration).