npx-based project creation
This solves a bug with the current TypeScript template caused by
package-lock.json not containing the additional packages
listed in the template's package.json, thus ending up removing
those packages the next time you run `npm install` on the newly created
project.

It also uses the new `npx` command, which reduces the size of the
command.