Remove Node 4 support
When using versions of NPM too old to understand `package-lock.json` (or stick to its rules, I'm not entirely sure), they will end up installing dependencies of dependencies of versions that are not consistent with our package-lock file.

That can in turn mean that node modules with known vulnerabilities get installed, and we would like to avoid that. For that reason I had to remove Node 4 from our test matrix. Not all our current dependency versions are ready for Node 9 yet, so I have not been able to add Node 9 to the matrix.

That means our test matrix will become:

- Node 6
- Node 7
- Node 8