mage.core.State: this === mage.core if new missing
Consider the following code:

```typescript
const state = mage.core.State();
```

While some linters will pick up on the mistake, others may not, and
the REPL interface will give you no hint that you are doing is wrong.

And what you are doing, is, indeed, terribly wrong:

- Instead of a new instance `this` in State is now equal to
  mage.core
- When we instanciate the object, we create a new Archivist instance:
  this.archivist = new Archivist()
- Which then replaces mage.core.archivist with an **instance** of
  Archivist; from there onward, archivist modules are no longer
  available.
- If you try to properly shut down or CTRL-C, missing API will create an
  emergency shutdown error; the process is now stuck.

JavaScript. Not even once.

The solution is to:

- Make sure `this` is defined
- Make sure `this.archivist` is NOT defined

It's mostly a band-aid, but hopefully this will avoid the majority of
the issues.